+++ winutil.ps1
@@ -10,7 +10,7 @@
     Author         : Chris Titus @christitustech
     Runspace Author: @DeveloperDurp
     GitHub         : https://github.com/ChrisTitusTech
-    Version        : 23.12.24
+    Version        : 23.12.30
 #>
 
 Start-Transcript $ENV:TEMP\Winutil.log -Append
@@ -22,7 +22,7 @@ Add-Type -AssemblyName System.Windows.Forms
 # Variable to sync between runspaces
 $sync = [Hashtable]::Synchronized(@{})
 $sync.PSScriptRoot = $PSScriptRoot
-$sync.version = "23.12.24"
+$sync.version = "23.12.30"
 $sync.configs = @{}
 $sync.ProcessRunning = $false
 
@@ -1135,15 +1135,22 @@ function Remove-ProvisionedPackages
 			$_.PackageName -NotLike "*Wifi*" -and
 			$_.PackageName -NotLike "*Foundation*" 
 		} 
-
-	$counter = 0
-	foreach ($appx in $appxProvisionedPackages)
-	{
-		$status = "Removing Provisioned $($appx.PackageName)"
-		Write-Progress -Activity "Removing Provisioned Apps" -Status $status -PercentComplete ($counter++/$appxProvisionedPackages.Count*100)
-		dism /image:$scratchDir /Remove-ProvisionedAppxPackage /PackageName:$($appx.PackageName) /NoRestart
-	}
-	Write-Progress -Activity "Removing Provisioned Apps" -Status "Ready" -Completed
+    
+    if ($?)
+    {
+	    $counter = 0
+	    foreach ($appx in $appxProvisionedPackages)
+	    {
+		    $status = "Removing Provisioned $($appx.PackageName)"
+		    Write-Progress -Activity "Removing Provisioned Apps" -Status $status -PercentComplete ($counter++/$appxProvisionedPackages.Count*100)
+		    dism /image:$scratchDir /Remove-ProvisionedAppxPackage /PackageName:$($appx.PackageName) /NoRestart
+	    }
+	    Write-Progress -Activity "Removing Provisioned Apps" -Status "Ready" -Completed
+    }
+    else
+    {
+        Write-Host "Could not get Provisioned App information. Skipping process..."
+    }
 }
 
 function Copy-ToUSB([string] $fileToCopy)
